+++ src/main/resources/templates/urls.py
@@ -1,0 +1,5 @@
+from django.urls import include, path, re_path
+
+urlpatterns = [
+    # Create your urls here.
+]